API Key and filesystem data access cleanup
I switched to using an environment variable (API_KEY_EIA) for access to
the EIA data. And also cleaned up some of the file storage in the local
data area.

Made a few notes in Markdown.

Could not get the notebook to run -- not 100% sure whether I changed
something or if it was not runnable from the beginning. It fails in the
sales_df area around cell 23.